Cost Function 16:26 11/22/2022
As opposed to just handing in the score the network earned and the desired score, a cost double is calculated by comparing each set of outputs to a set of expected outputs for each image.
MSE, or Mean Squared Error, is the function used for calculating the cost double.
Reference link: https://www.analyticsvidhya.com/blog/2021/02/cost-function-is-no-rocket-science/